requirements.txt
mysql-connector-python==8.0.29
flask
Flask-Bootstrap4
pyjwt